% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_ass.R
\name{orm_ass_plot}
\alias{orm_ass_plot}
\title{Plot ASS distribution}
\usage{
orm_ass_plot(mx, out_dir = NULL, lang = getOption("orisma.lang", "en"))
}
\arguments{
\item{mx}{An \code{orisma_matrix} object after running \code{\link[=orm_ass]{orm_ass()}}.}

\item{out_dir}{Character or NULL. Directory to save the plot.}

\item{lang}{Character. \code{"en"} or \code{"es"}.}
}
\value{
A ggplot2 object invisibly.
}
\description{
Generates a bar chart showing the distribution of Abstract Sufficiency
Scores across the corpus.
}
