% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_dict.R
\name{orm_dict}
\alias{orm_dict}
\title{Load a risk dictionary}
\usage{
orm_dict(name = "iso45001_insst")
}
\arguments{
\item{name}{Character. Dictionary name. Default \code{"iso45001_insst"}.}
}
\value{
A named list (class \code{orisma_dict}).
}
\description{
Load a risk dictionary
}
