% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_dict.R
\name{orm_dict_add_terms}
\alias{orm_dict_add_terms}
\title{Add terms to an existing dictionary category}
\usage{
orm_dict_add_terms(dict, category, terms)
}
\arguments{
\item{dict}{An \code{orisma_dict} object.}

\item{category}{Character. Category key.}

\item{terms}{Character vector. New terms to add.}
}
\value{
Updated \code{orisma_dict}.
}
\description{
Add terms to an existing dictionary category
}
