% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_dict.R
\name{print.orisma_dict}
\alias{print.orisma_dict}
\title{Print method for orisma_dict}
\usage{
\method{print}{orisma_dict}(x, ...)
}
\arguments{
\item{x}{An \code{orisma_dict} object.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for orisma_dict
}
