% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_autodim.R
\name{print.orisma_dims}
\alias{print.orisma_dims}
\title{Print method for orisma_dims}
\usage{
\method{print}{orisma_dims}(x, ...)
}
\arguments{
\item{x}{An \code{orisma_dims} object.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for orisma_dims
}
