% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_validate.R
\name{print.orisma_kappa}
\alias{print.orisma_kappa}
\title{Print method for orisma_kappa}
\usage{
\method{print}{orisma_kappa}(x, ...)
}
\arguments{
\item{x}{An \code{orisma_kappa} object.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for orisma_kappa
}
