% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_extract.R
\name{print.orisma_matrix}
\alias{print.orisma_matrix}
\title{Print method for orisma_matrix}
\usage{
\method{print}{orisma_matrix}(x, ...)
}
\arguments{
\item{x}{An object to print.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for orisma_matrix
}
