% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_normativa.R
\name{print.orisma_normativa}
\alias{print.orisma_normativa}
\title{Print method for orisma_normativa}
\usage{
\method{print}{orisma_normativa}(x, ...)
}
\arguments{
\item{x}{An \code{orisma_normativa} object.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for orisma_normativa
}
