% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_priority.R
\name{print.orisma_priority}
\alias{print.orisma_priority}
\title{Print method for orisma_priority}
\usage{
\method{print}{orisma_priority}(x, ...)
}
\arguments{
\item{x}{An \code{orisma_priority} object.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for orisma_priority
}
