% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_bridge.R
\name{print.orisma_ranking}
\alias{print.orisma_ranking}
\title{Print method for orisma_ranking}
\usage{
\method{print}{orisma_ranking}(x, ...)
}
\arguments{
\item{x}{An \code{orisma_ranking} object.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for orisma_ranking
}
