% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_analyse.R
\name{print.orisma_result}
\alias{print.orisma_result}
\title{Print method for orisma_result}
\usage{
\method{print}{orisma_result}(x, ...)
}
\arguments{
\item{x}{An object to print.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for orisma_result
}
