% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orisma-package.R
\docType{package}
\name{orisma-package}
\alias{orisma-package}
\alias{orisma}
\title{orisma: Occupational Risk Integrated Systematic Mapping and Analysis}
\description{
A systematic bibliometric pipeline for mapping occupational risk evidence
in any domain - emerging technologies, established hazards, or specific
industrial sectors.
\subsection{Typical workflow}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(orisma)
data   <- orm_load("my_references/")
result <- orm_run(data)
orm_report(result)
}\if{html}{\out{</div>}}
}
}
\section{Original indicators}{

\itemize{
\item \strong{WRDI} Worker-Risk Disconnection Index
\item \strong{RCS} Risk Category Saturation Index
\item \strong{MGP} Material-Gap Profile
}
}

\section{Citation}{

Aguilar-Elena, R. (2025). orisma: Occupational Risk Integrated Systematic
Mapping and Analysis. R package version 0.1.0.
Universidad Internacional de Valencia (VIU).
\url{https://github.com/Aguilar-Elena/orisma}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Aguilar-Elena/orisma}
  \item Report bugs at \url{https://github.com/Aguilar-Elena/orisma/issues}
}

}
\author{
Raul Aguilar-Elena \email{raguilar@universidadviu.com}

Occupational Risk Prevention and Occupational Health Research Group (GPRL),
Universidad Internacional de Valencia (VIU), Valencia, Spain.
}
