% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_dict.R
\name{orm_dict_categories}
\alias{orm_dict_categories}
\title{List risk categories in a dictionary}
\usage{
orm_dict_categories(dict, lang = getOption("orisma.lang", "en"))
}
\arguments{
\item{dict}{An \code{orisma_dict} object.}

\item{lang}{Character. \code{"en"} or \code{"es"}.}
}
\description{
List risk categories in a dictionary
}
