% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orm_dict.R
\name{orm_dict_list}
\alias{orm_dict_list}
\title{List available built-in dictionaries}
\usage{
orm_dict_list()
}
\value{
A character vector with the names of the built-in dictionaries available in ORISMA.
}
\description{
List available built-in dictionaries
}
\details{
This function takes no arguments.
}
