% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orisma-package.R, R/orm_extract.R
\name{print.orisma_matrix}
\alias{print.orisma_matrix}
\title{Print an orisma_matrix object}
\usage{
\method{print}{orisma_matrix}(x, ...)

\method{print}{orisma_matrix}(x, ...)
}
\arguments{
\item{x}{An \code{orisma_matrix} object.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an orisma_matrix object

Print method for orisma_matrix
}
