% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orisma-package.R, R/orm_analyse.R
\name{print.orisma_result}
\alias{print.orisma_result}
\title{Print an orisma_result object}
\usage{
\method{print}{orisma_result}(x, ...)

\method{print}{orisma_result}(x, ...)
}
\arguments{
\item{x}{An \code{orisma_result} object.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an orisma_result object

Print method for orisma_result
}
